+++ tests/testthat/testthat_3d.r
@@ -12,10 +12,11 @@ Srs3 = Polygons(list(Sr3, Sr4), "s3/4")
 Sp <- SpatialPolygonsDataFrame( SpatialPolygons(list(Srs1,Srs2,Srs3), 1:3), data.frame(x = 1:3, row.names = c("s1", "s2", "s3/4")))
 
 context("test 3d")
-
+x0 <- gris(Sp)
+x1 <- triangulate(x0)
 test_that("we can triangulate", {
-  expect_that(x1 <- gris(Sp), is_a("gris"))
-  expect_that(x1 <- triangulate(x1), is_a("gris"))
+  expect_that(x0, is_a("gris"))
+  expect_that(x1, is_a("gris"))
   expect_that(x1$tXv, is_a("tbl"))
   expect_that(x1$oXt, is_a("tbl"))
   
@@ -32,6 +33,6 @@ test_that("we can triangulate", {
 
 library(rgl)
 test_that("we can plot without error", {
-  expect_silent(plot3d(x1, globe = FALSE))
+  expect_silent(gris::plot3d(x1, globe = FALSE))
 })